+++ src/hkimjp/wil2/admin.clj
@@ -11,9 +11,9 @@
     [:div.text-2xl "admin"]
     [:div.m-4
      [:table
-      [:tr [:td.font-bold "develop"] [:td (env :develop)]]
-      [:tr [:td.font-bold "auth"] [:td (env :auth)]]
-      [:tr [:td.font-bold "datascript"] [:td (env :datascript)]]
-      [:tr [:td.font-bold "redis"] [:td (env :redis)]]
-      [:tr [:td.font-bold "max-count"] [:td max-count]]
-      [:tr [:td.font-bold "min-interval　"] [:td min-interval]]]]]))
+      [:tr [:td.font-bold "develop"] [:td] [:td (env :develop)]]
+      [:tr [:td.font-bold "auth"] [:td] [:td (env :auth)]]
+      [:tr [:td.font-bold "datascript"] [:td] [:td (env :datascript)]]
+      [:tr [:td.font-bold "redis"] [:td] [:td (env :redis)]]
+      [:tr [:td.font-bold "max-count"] [:td] [:td max-count]]
+      [:tr [:td.font-bold "min-interval"] [:td.w-4 ""] [:td min-interval]]]]]))