+++ src/hkimjp/wil2/points.clj
@@ -54,7 +54,7 @@
     (t/log! :info (str "my " user "recv " recv))
     (page
      [:div.mx-4
-      [:div.text-2xl user "'s points"]
+      [:div.text-2xl "Points: " user]
       [:p "送信ポイント、受信ポイントの和が平常点。"]
       [:div
        [:div.font-bold.py-2 "送信ポイント"]
@@ -71,8 +71,7 @@
           (format "⬆️ %d,  ➡️ %d, ⬇️ %d == %d" a b c (+ (* 2 a) b (* -1 c))))]
        [:div.font-bold.py-4 "自分の WIL と獲得ポイント"]
        [:p "自分が提出した WIL の下に"
-        [:span.text-red-600 "獲得ポイント"]
-        "を表示している。"]
+        [:span.text-red-600 "獲得ポイント"] "。"]
        [:div
         (for [[e _ md] (sort-by second (ds/qq my-uploads user))]
           (conj (-> md