+++ src/hkimjp/wil2/switch.clj
@@ -36,15 +36,16 @@
       [:p.py-2 "授業日中に今日の WIL を提出、授業後3日以内に他受講生の WIL を評価する。"]
       (if develop?
         [:ul
-         [:li [:a.hover:underline {:href "/wil2/upload"} "今日のWILを提出"]]
-         [:li [:a.hover:underline {:href "/wil2/rating"} "今週のWILを評価"]]]
+         [:li.py-2 [:a.hover:underline {:href "/wil2/upload"} "今日のWILを提出"]]
+         [:li.py-2 [:a.hover:underline {:href "/wil2/rating"} "今週のWILを評価"]]]
         [:ul
-         [:li (cond
-                (not (can-upload?)) [:span "WIL が提出できるのは授業のあった日。"]
-                uploaded?           [:span "提出済みです。"]
-                :else               [:a.hover:underline
-                                     {:href "/wil2/upload"} "今日のWILを提出"])]
-         [:li (if (can-rate?)
-                [:a.hover:underline {:href "/wil2/rating"} "今週のWILを評価"]
-                [:span "rating 期間終了。"])]])])))
+         [:li.py-2 (cond
+                     (not (can-upload?)) [:span "WIL が提出できるのは授業のあった日。"]
+                     uploaded?           [:span "提出済みです。"]
+                     :else               [:a.hover:underline
+                                          {:href "/wil2/upload"} "今日のWILを提出"])]
+         [:li.py-2 (if (can-rate?)
+                     [:a.hover:underline {:href "/wil2/rating"} "今週のWILを評価"]
+                     [:span "rating 期間終了。"])]])
+      [:br]])))